pyproject.toml
[project]
name = "flask-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "psycopg2-binary>=2.9.11",
]
